Restructuration: negfc and metrics routines that are fm-related were placed in a 'fm' module
