+++ setup.py
@@ -60,9 +60,9 @@ with open(resource('vip_hci', '__init__.py')) as version_file:
 PACKAGES = ['vip_hci',
             'vip_hci.config',
             'vip_hci.fits',
+            'vip_hci.fm',
             'vip_hci.invprob',
             'vip_hci.metrics',
-            'vip_hci.negfc',
             'vip_hci.preproc',
             'vip_hci.psfsub',
             'vip_hci.stats',
