+++ vip_hci/fm/negfc_speckle_noise.py
