+++ vip_hci/fm/utils_negfc.py
