+++ vip_hci/metrics/contrcurve.py
@@ -19,8 +19,8 @@ from scipy import stats
 from scipy.signal import savgol_filter
 from skimage.draw import disk
 from matplotlib import pyplot as plt
-from .fakecomp import (cube_inject_companions, frame_inject_companion,
-                       normalize_psf)
+from ..fm import (cube_inject_companions, frame_inject_companion,
+                  normalize_psf)
 from ..config import time_ini, timing
 from ..config.utils_conf import sep
 from ..var import frame_center, dist
